refactor: move shader code for clarity
