+++ cryoet_data_portal_neuroglancer/models/json_generator.py
@@ -5,7 +5,7 @@ from typing import Any
 
 import numpy as np
 
-from cryoet_data_portal_neuroglancer.models.shader_builder import ImageShaderBuilder
+from cryoet_data_portal_neuroglancer.shaders.shader_builder import ImageShaderBuilder
 
 
 def create_source(
